+++ src/main/java/com/programmers/board/controller/PostController.java
@@ -4,12 +4,15 @@ import com.programmers.board.constant.SessionConst;
 import com.programmers.board.constant.AuthErrorMessage;
 import com.programmers.board.controller.response.PageResult;
 import com.programmers.board.controller.response.Result;
-import com.programmers.board.dto.PostDto;
-import com.programmers.board.dto.request.PostCreateRequest;
-import com.programmers.board.dto.request.PostUpdateRequest;
-import com.programmers.board.dto.service.post.*;
+import com.programmers.board.service.response.PostDto;
+import com.programmers.board.controller.request.PostCreateRequest;
+import com.programmers.board.controller.request.PostUpdateRequest;
 import com.programmers.board.exception.AuthenticationException;
 import com.programmers.board.service.PostService;
+import com.programmers.board.service.request.post.PostCreateCommand;
+import com.programmers.board.service.request.post.PostDeleteCommand;
+import com.programmers.board.service.request.post.PostGetCommand;
+import com.programmers.board.service.request.post.PostUpdateCommand;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
