+++ src/main/java/com/programmers/board/controller/UserController.java
@@ -4,13 +4,13 @@ import com.programmers.board.constant.SessionConst;
 import com.programmers.board.constant.AuthErrorMessage;
 import com.programmers.board.controller.response.PageResult;
 import com.programmers.board.controller.response.Result;
-import com.programmers.board.dto.UserDto;
-import com.programmers.board.dto.request.UserCreateRequest;
-import com.programmers.board.dto.request.UserUpdateRequest;
-import com.programmers.board.dto.service.user.UserCreateCommand;
-import com.programmers.board.dto.service.user.UserDeleteCommand;
-import com.programmers.board.dto.service.user.UserGetCommand;
-import com.programmers.board.dto.service.user.UserUpdateCommand;
+import com.programmers.board.service.response.UserDto;
+import com.programmers.board.controller.request.UserCreateRequest;
+import com.programmers.board.controller.request.UserUpdateRequest;
+import com.programmers.board.service.request.user.UserCreateCommand;
+import com.programmers.board.service.request.user.UserDeleteCommand;
+import com.programmers.board.service.request.user.UserGetCommand;
+import com.programmers.board.service.request.user.UserUpdateCommand;
 import com.programmers.board.exception.AuthenticationException;
 import com.programmers.board.service.UserService;
 import jakarta.validation.Valid;
