+++ src/main/java/com/programmers/board/controller/request/PostUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.programmers.board.dto.request;
+package com.programmers.board.controller.request;
 
 import jakarta.validation.constraints.Size;
 import lombok.AllArgsConstructor;
