+++ src/main/java/com/programmers/board/controller/request/UserCreateRequest.java
@@ -1,4 +1,4 @@
-package com.programmers.board.dto.request;
+package com.programmers.board.controller.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
