+++ src/main/java/com/programmers/board/service/PostService.java
@@ -3,11 +3,14 @@ package com.programmers.board.service;
 import com.programmers.board.constant.AuthErrorMessage;
 import com.programmers.board.domain.Post;
 import com.programmers.board.domain.User;
-import com.programmers.board.dto.PostDto;
-import com.programmers.board.dto.service.post.*;
+import com.programmers.board.service.response.PostDto;
 import com.programmers.board.exception.AuthorizationException;
 import com.programmers.board.repository.PostRepository;
 import com.programmers.board.repository.UserRepository;
+import com.programmers.board.service.request.post.PostCreateCommand;
+import com.programmers.board.service.request.post.PostDeleteCommand;
+import com.programmers.board.service.request.post.PostGetCommand;
+import com.programmers.board.service.request.post.PostUpdateCommand;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
