+++ src/main/java/com/programmers/board/service/request/post/PostGetCommand.java
@@ -1,4 +1,4 @@
-package com.programmers.board.dto.service.post;
+package com.programmers.board.service.request.post;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
