+++ src/test/java/com/programmers/board/controller/PostControllerTest.java
@@ -4,9 +4,9 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.programmers.board.constant.SessionConst;
 import com.programmers.board.domain.Post;
 import com.programmers.board.domain.User;
-import com.programmers.board.dto.PostDto;
-import com.programmers.board.dto.request.PostCreateRequest;
-import com.programmers.board.dto.request.PostUpdateRequest;
+import com.programmers.board.service.response.PostDto;
+import com.programmers.board.controller.request.PostCreateRequest;
+import com.programmers.board.controller.request.PostUpdateRequest;
 import com.programmers.board.service.PostService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
