+++ src/test/java/com/programmers/board/controller/UserControllerTest.java
@@ -3,9 +3,9 @@ package com.programmers.board.controller;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.programmers.board.constant.SessionConst;
 import com.programmers.board.domain.User;
-import com.programmers.board.dto.UserDto;
-import com.programmers.board.dto.request.UserCreateRequest;
-import com.programmers.board.dto.request.UserUpdateRequest;
+import com.programmers.board.service.response.UserDto;
+import com.programmers.board.controller.request.UserCreateRequest;
+import com.programmers.board.controller.request.UserUpdateRequest;
 import com.programmers.board.service.UserService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
